+++ examples/realworld.svelte.dev/src/components/ListErrors.svelte
@@ -1,0 +1,11 @@
+<script>
+	export let errors;
+</script>
+
+{#if errors}
+	<ul class="error-messages">
+		{#each Object.keys(errors) as key}
+			<li>{key} {errors[key]}</li>
+		{/each}
+	</ul>
+{/if}