+++ examples/realworld.svelte.dev/src/routes/$error.svelte
@@ -1,0 +1,46 @@
+<script>
+	import Layout from './$layout.svelte';
+
+	export let error, status;
+
+	let dev = import.meta.env.MODE === 'development';
+</script>
+
+<svelte:head>
+	<title>{status}</title>
+</svelte:head>
+
+<Layout>
+	<div class="col-md-9">
+		<h1>{status}</h1>
+
+		<p>{error.message}</p>
+
+		{#if dev && error.stack}
+			<pre>{error.stack}</pre>
+		{/if}
+	</div>
+</Layout>
+
+<style>
+	h1,
+	p {
+		margin: 0 auto;
+	}
+
+	h1 {
+		font-size: 2.8em;
+		font-weight: 700;
+		margin: 0 0 0.5em 0;
+	}
+
+	p {
+		margin: 1em auto;
+	}
+
+	@media (min-width: 480px) {
+		h1 {
+			font-size: 4em;
+		}
+	}
+</style>