+++ examples/realworld.svelte.dev/src/routes/$layout.svelte
@@ -1,0 +1,12 @@
+<script>
+	import { stores } from '$app/client.js';
+	import Nav from '$components/Nav.svelte';
+
+	export let segment;
+</script>
+
+<Nav {segment}/>
+
+<main>
+	<slot></slot>
+</main>