+++ examples/realworld.svelte.dev/src/routes/auth/logout.js
@@ -1,0 +1,4 @@
+export function post(req, res) {
+	delete req.session.user;
+	res.end(JSON.stringify({ ok: true }));
+}