+++ examples/realworld.svelte.dev/src/routes/editor/index.svelte
@@ -1,0 +1,14 @@
+<script context="module">
+	export function preload({ params }, { user }) {
+		if (!user) {
+			this.redirect(302, `/login`);
+		}
+	}
+</script>
+
+<script>
+	import Editor from './_Editor.svelte';
+	let article = { title: '', description: '', body: '', tagList: [] };
+</script>
+
+<Editor {article}/>