+++ examples/realworld.svelte.dev/src/setup/index.js
@@ -1,0 +1,12 @@
+export function prepare(headers) {
+	// TODO
+	return {
+		context: {},
+		headers: {}
+	};
+}
+
+export function getSession(context) {
+	// TODO
+	return context;
+}